project2-FaceRecognition/requirements.txt
numpy
matplotlib
argparse
logging
scikit-learn
pillow